+++ src/routes/messages/frontend_models.rs
@@ -1,5 +1,5 @@
 use crate::common::models::addresses::AddressEx;
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use std::collections::BTreeMap;
 
 #[derive(Serialize, Debug)]
@@ -17,16 +17,6 @@ pub(super) struct Confirmation {
 }
 
 #[derive(Serialize, Debug)]
-#[serde(tag = "type", rename_all = "SCREAMING_SNAKE_CASE")]
-pub(super) enum MessageItem {
-    Message(Message),
-    #[serde(rename_all = "camelCase")]
-    DateLabel {
-        timestamp: i64,
-    },
-}
-
-#[derive(Serialize, Debug)]
 #[serde(rename_all = "camelCase")]
 pub(super) struct Message {
     pub(super) message_hash: String,
@@ -41,20 +31,6 @@ pub(super) struct Message {
     pub(super) proposed_by: AddressEx,
     pub(super) confirmations: Vec<Confirmation>,
     pub(super) prepared_signature: Option<String>,
-}
-
-#[derive(Deserialize, Serialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct CreateMessage {
-    message: String,
-    safe_app_id: Option<u64>,
-    signature: String,
-}
-
-#[derive(Deserialize, Serialize, Debug)]
-#[serde(rename_all = "camelCase")]
-pub struct UpdateMessage {
-    signature: String,
 }
 
 #[derive(Serialize, Debug)]
